% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.KW2.R
\name{bounds.KW2}
\alias{bounds.KW2}
\title{Prediction of Bounds on Marginal Effects}
\usage{
bounds.KW2(object, ...)
}
\arguments{
\item{object}{is the fitted NPMLE object}

\item{\dots}{is expected to contain an argument \code{newdata}}
}
\value{
a list consisting of the following components:
\describe{
	\item{phat}{Point prediction}
	\item{lower}{lower bound prediction}
	\item{upper}{upper bound prediction}
	\item{xpoly}{indices of crossed polygons}
}
}
\description{
Given a fitted model by the exact NPMLE procedure prediction is made
at a new design point with lower and upper bounds for the prediction due
to ambiguity of the assignment of mass within the cell enumerated polygons.
}
\seealso{
\code{predict.KW2} for a simpler prediction function without bounds
}
\author{
Jiaying Gu
}
