% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GH.R
\name{GH.se}
\alias{GH.se}
\title{Current Status Linear Regression Standard Errors}
\usage{
GH.se(bstar, X, y, eps = 0.001, hc = 2)
}
\arguments{
\item{bstar}{parameter vector (fix last entry as a known number, usually 1 or -1, for normalization)}

\item{X}{design matrix}

\item{y}{binary response vector}

\item{eps}{trimming tolerance parameter}

\item{hc}{kernel bandwidth (used for the standard error estimation)}
}
\value{
A list with components:
\itemize{
\item  evaluation of a score function at parameter value
\item  estimated standard error 
\item  sindex single index linear predictor
}
}
\description{
Groeneboom and Hendrickx semiparametric binary response estimator (scalar case)
score estimator based on NPMLE avoids any smoothing
proposed by Groneboom and Hendrickx (2018).
}
\references{
Groeneboom, P. and K. Hendrickx (2018) Current Status Linear Regression, 
 Annals of Statistics, 46, 1415-1444,
}
