% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Horowitz93.R
\docType{data}
\name{Horowitz93}
\alias{Horowitz93}
\title{Horowitz (1993) Modal Choice Data}
\format{
A data frame with 842 observations on 5 variables:
}
\source{
\url{https://www.gams.com/latest/gamslib_ml/libhtml/gamslib_mws.html}
}
\usage{
Horowitz93
}
\description{
Modal choice data for journey to work in the Washington DC area
from  the late 1960's.  The variables are:
* `DCOST`: difference in cost of car versus transit (transit - car)
* `CARS`: number of cars at home 
* `DOVTT`: difference in out of vehicle time (transit - car)
* `DIVTT`: difference in in vehicle time (transit - car)
* `DEPEND`: coded 1 if by car, 0 if by mass transit
}
\references{
Horowitz, J L, (1993) Semiparametric estimation of a work-trip mode choice model.
Journal of Econometrics, 58, 49-70.
}
\keyword{datasets}
