% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.GK.R
\name{logLik.GK}
\alias{logLik.GK}
\title{log likelihood for Gautier Kitamura procedure}
\usage{
\method{logLik}{GK}(object, ...)
}
\arguments{
\item{object}{a fitted object of class "GK"}

\item{...}{other parameters for logLik}
}
\value{
a scalar log likelihood
}
\description{
log likelihood for Gautier Kitamura procedure
}
