% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.KW1.R
\name{logLik.KW1}
\alias{logLik.KW1}
\title{log likelihood for KW1 procedure}
\usage{
\method{logLik}{KW1}(object, ...)
}
\arguments{
\item{object}{a fitted object of class "KW1"}

\item{...}{other parameters for logLik}
}
\value{
a scalar log likelihood
}
\description{
log likelihood for KW1 procedure
}
