% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbours.R
\name{neighbours}
\alias{neighbours}
\title{Check Neighbouring Cell Counts}
\usage{
neighbours(SignVector)
}
\arguments{
\item{SignVector}{n by m matrix of signs produced by NICER}
}
\value{
Column indices of the cells that are locally maximal, i.e. those
whose neighbours have strictly fewer cell counts.  The corresponding
interior points of these cells can be used as potential mass points
for the NPMLE function \code{rcbr.fit.KW}.
}
\description{
Compare cell counts for each cell with its neighbours and return indices
of the locally maximal cells.
}
