% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GK.R
\name{plot.GK}
\alias{plot.GK}
\title{Plot a GK object}
\usage{
\method{plot}{GK}(x, ...)
}
\arguments{
\item{x}{is the fitted GK object}

\item{...}{other arguments to pass to \code{contour}, notably e.g. \code{add = TRUE}}
}
\value{
nothing (invisibly)
}
\description{
Given a fitted model by the Guatier-Kitamura procedure plot the estimated density contours
}
