% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.KW2.R
\name{plot.KW2}
\alias{plot.KW2}
\title{Plot a KW2 object}
\usage{
\method{plot}{KW2}(x, smooth = 0, pal = NULL, inches = 1/6, N = 25, tol = 0.001, ...)
}
\arguments{
\item{x}{is the fitted NPMLE object}

\item{smooth}{is a parameter to control bandwidth of the smoothing if a contour plot
of the estimated density is desired, default is no smoothing and only
the mass points of the discrete estimate are plotted.}

\item{pal}{a color palette}

\item{inches}{as used in \code{symbols} to control size of mass points}

\item{N}{scaling of the color palette}

\item{tol}{tolerance for size of mass points}

\item{...}{other arguments to pass to \code{symbols}, notably e.g. \code{add = TRUE}}
}
\value{
nothing (invisibly)
}
\description{
Given a fitted model by the rcbr NPMLE procedure plot the estimated mass points
}
