% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyzone.R
\name{polyzone}
\alias{polyzone}
\title{Identify crossed polygons from existing cells when adding a new line (works only for dim = 2)}
\usage{
polyzone(SignVector, w, A, b)
}
\arguments{
\item{SignVector}{current SignvVctor matrix}

\item{w}{associated interior points}

\item{A}{design matrix for full problem aka [1,z]}

\item{b}{associated final column of design matrix aka [v]}
}
\value{
vector of indices of crossed polygons
}
\description{
Given an existing cell configuration represented by the Signvector and
associated interior points w, identify the polygons crossed by the next
new line.
}
\author{
Jiaying Gu
}
