% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.GK.R
\name{predict.GK}
\alias{predict.GK}
\title{Prediction of Marginal Effects}
\usage{
\method{predict}{GK}(object, ...)
}
\arguments{
\item{object}{is the fitted object of class "GK"}

\item{\dots}{is expected to contain an argument \code{newdata}}
}
\value{
a vector pf predicted probabilities
}
\description{
Given a fitted model by the Gautier Kitamura procedure predictions are made
at new design points given by the \code{newdata} argument.
}
