% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.KW2.R
\name{predict.KW2}
\alias{predict.KW2}
\title{Prediction of Marginal Effects}
\usage{
\method{predict}{KW2}(object, ...)
}
\arguments{
\item{object}{is the fitted NPMLE object}

\item{\dots}{is expected to contain an argument \code{newdata}}
}
\value{
a vector pf predicted probabilities
}
\description{
Given a fitted model by the rcbr NPMLE procedure predictions are made
at new design points given by the \code{newdata} argument.
}
\seealso{
\code{bound.KW2} for a prediction function with bounds
}
